+++ router/page.ts
@@ -1,26 +1,12 @@
 import {
   Middleware,
-  Params,
 } from './type.ts'
-import { pageTupples } from './router.ts'
+import { router } from './router.ts'
 import { open } from './open.ts'
 
 export function page(pattern: string, ...middlewares: Middleware[]): void
 {
-  let names = pattern.split('/')
-  let len = names.length
-  while (pageTupples.length < len + 1) {
-    pageTupples.push([new Set(), new Map()])
-  }
-  let kinds = pageTupples[len][0]
-  let pages = pageTupples[len][1]
-  let kind = names.reduce((kind, name, index) => kind | ((name[0] === ':' ? 1 : 0) << (len - 1 - index)), 0)
-  let key = names.map(name => name[0] === ':' ? '*' : name).join('/')
-  let params: Params = []
-  names.forEach((name, index) => name[0] === ':' && params.push([name.slice(1), index]))
-  kinds.add(kind)
-  pageTupples[len][0] = new Set(Array.from(kinds).sort())
-  pages.set(key, [pattern, params, middlewares])
+  router.page(pattern, ...middlewares)
 }
 
 // export function page(pattern: string, component: Component, elementize?: Elementize): void
